Description:add get uint64_t tokenId interface Feature or Bugfix:Feature Binary Source: No

+++ ipc/native/src/mock/include/iremote_invoker.h
@@ -66,9 +66,9 @@ public:
 
     virtual uid_t GetCallerUid() const = 0;
 
-    virtual uint32_t GetCallerTokenID() const = 0;
+    virtual uint64_t GetCallerTokenID() const = 0;
 
-    virtual uint32_t GetFirstTokenID() const = 0;
+    virtual uint64_t GetFirstTokenID() const = 0;
 
     virtual uint32_t GetStatus() const = 0;
